+++ src/middlewares/connectTo.ts
@@ -1,6 +1,0 @@
-import {global} from "../tools/global";
-
-const connectToXXX = (): void => {
-  global._XXX_ = (store: any) => (next: any) => (action: string) => {
-  };
-};