+++ tests/utils/hideConsoleError.ts
@@ -1,8 +1,0 @@
-const consoleError = console.error;
-
-export const hideConsoleError = (hide: boolean): void => {
-  hide
-    ? console.error = () => undefined
-    : console.error = consoleError;
-};
-