+++ doc/Change-Log.md
@@ -116,6 +116,10 @@ This allows actions to be used as events.
 
 There was a bug where if an implementation of an _overloaded_ action return no change, the change wasn't applied.  
 
+### v3.x
+
+- `IDynaduxConfig.middlewares` can be only `IDynaduxMiddleware<any, any>[]`, not `false`, `null` or `undefined` anymore
+- Remove the `TPayload` from interfaces, needed only for the dispatch method
 
 # Read more